feat: responsive redesign of Teachers
Split 685-line file into Teachers (281), TeacherFormModal (294),
DeleteTeacherModal (60). Remove palette object, use UI components,
add responsive table columns with mobile sub-info in name cell.
