+++ front/src/pages/teachers/DeleteTeacherModal.tsx
@@ -1,0 +1,60 @@
+import { Trash2 } from "lucide-react";
+import { Button } from "@/components/ui/button";
+
+export interface DeleteTeacherModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+  onConfirm: () => void;
+  professorName: string;
+}
+
+export function DeleteTeacherModal({
+  isOpen,
+  onClose,
+  onConfirm,
+  professorName,
+}: DeleteTeacherModalProps) {
+  if (!isOpen) return null;
+
+  return (
+    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/30 backdrop-blur-sm p-4">
+      <div className="w-full max-w-md rounded-xl border border-stone-200 bg-white p-4 shadow-lg sm:p-6">
+        <div className="mb-4 flex items-center gap-3">
+          <div className="rounded-lg bg-red-50 p-2">
+            <Trash2 className="h-6 w-6 text-red-600" />
+          </div>
+          <h2 className="text-xl font-bold text-stone-900">
+            Confirmar Remoção
+          </h2>
+        </div>
+
+        <p className="mb-6 text-sm text-stone-500 sm:text-base">
+          Tem certeza que deseja remover o professor{" "}
+          <span className="font-semibold text-stone-900">
+            {professorName}
+          </span>
+          ? Esta ação não pode ser desfeita.
+        </p>
+
+        <div className="flex gap-3">
+          <Button
+            type="button"
+            variant="outline"
+            onClick={onClose}
+            className="flex-1 rounded-xl border-stone-300 text-stone-700 hover:bg-stone-50"
+          >
+            Cancelar
+          </Button>
+          <Button
+            type="button"
+            variant="destructive"
+            onClick={onConfirm}
+            className="flex-1 rounded-xl"
+          >
+            Confirmar Remoção
+          </Button>
+        </div>
+      </div>
+    </div>
+  );
+}
